requirements.txt
Pillow
einops
safetensors
timm
tomesd
torchdiffeq
torchsde
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
onnxruntime
flask
flask_cors
accelerate>=1.10.1
gradio>=3.41.2
diffusers>=0.30.1
transformers==4.57.0
kornia==0.8.0